requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langgraph>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
typing-extensions>=4.0.0

# Web Interface
streamlit>=1.28.0

# Search Integration
langchain_community>=0.1.0
google-search-results>=2.4.0

# Vector Database & Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
langchain-chroma>=0.1.0

# Document Processing
pypdf>=5.0.0
python-docx>=1.0.0
unstructured>=0.15.0

# Additional Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
colorama>=0.4.0
